+++ packages/@office-iss/react-native-win32-tester/overrides.json
@@ -5,7 +5,7 @@
   "excludePatterns": [
     "src/js/examples-win32/**"
   ],
-  "baseVersion": "0.0.0-d40cb0e1b",
+  "baseVersion": "0.0.0-fa768df2a",
   "overrides": [
     {
       "type": "patch",
@@ -41,7 +41,7 @@
       "type": "derived",
       "file": "src/js/utils/RNTesterList.win32.js",
       "baseFile": "packages/rn-tester/js/utils/RNTesterList.android.js",
-      "baseHash": "cab581fd600f109a2efae881734e19e07c3bd350"
+      "baseHash": "ddd1efab44c0998488de268c691aeec9ad063f22"
     },
     {
       "type": "patch",
