+++ packages/@office-iss/react-native-win32-tester/package.json
@@ -16,7 +16,7 @@
   "peerDependencies": {
     "@office-iss/react-native-win32": "^0.0.0-canary.112",
     "react": "17.0.2",
-    "react-native": "0.0.0-d40cb0e1b"
+    "react-native": "0.0.0-fa768df2a"
   },
   "devDependencies": {
     "@office-iss/react-native-win32": "^0.0.0-canary.112",
@@ -26,7 +26,7 @@
     "@types/node": "^14.14.22",
     "eslint": "7.12.0",
     "just-scripts": "^1.3.3",
-    "react-native": "0.0.0-d40cb0e1b",
+    "react-native": "0.0.0-fa768df2a",
     "react-native-platform-override": "^1.4.18",
     "typescript": "^3.8.3"
   },
