+++ packages/@office-iss/react-native-win32/package.json
@@ -70,14 +70,14 @@
     "just-scripts": "^1.3.3",
     "prettier": "1.19.1",
     "react": "17.0.2",
-    "react-native": "0.0.0-d40cb0e1b",
+    "react-native": "0.0.0-fa768df2a",
     "react-native-platform-override": "^1.4.18",
     "react-shallow-renderer": "16.14.1",
     "typescript": "^3.8.3"
   },
   "peerDependencies": {
     "react": "17.0.2",
-    "react-native": "0.0.0-d40cb0e1b"
+    "react-native": "0.0.0-fa768df2a"
   },
   "beachball": {
     "defaultNpmTag": "canary",
