+++ packages/@react-native-windows/tester/package.json
@@ -16,7 +16,7 @@
   "peerDependencies": {
     "@react-native-picker/picker": "1.16.5",
     "react": "17.0.2",
-    "react-native": "0.0.0-d40cb0e1b",
+    "react-native": "0.0.0-fa768df2a",
     "react-native-windows": "^0.0.0-canary.382"
   },
   "devDependencies": {
@@ -26,7 +26,7 @@
     "@types/node": "^14.14.22",
     "eslint": "7.12.0",
     "just-scripts": "^1.3.3",
-    "react-native": "0.0.0-d40cb0e1b",
+    "react-native": "0.0.0-fa768df2a",
     "react-native-platform-override": "^1.4.18",
     "react-native-windows": "^0.0.0-canary.382",
     "typescript": "^3.8.3"
