+++ packages/@react-native/repo-config/overrides.json
@@ -1,5 +1,5 @@
 {
-  "baseVersion": "0.0.0-d40cb0e1b",
+  "baseVersion": "0.0.0-fa768df2a",
   "overrides": [
     {
       "type": "copy",
