+++ packages/@react-native/tester/js/examples/Animated/MovingBoxExample.js
@@ -78,7 +78,7 @@ function MovingBoxView({useNativeDriver}: {useNativeDriver: boolean}) {
   };
   return (
     <View style={styles.container}>
-      <View style={styles.boxContainer}>
+      <View testID="moving-view-track" style={styles.boxContainer}>
         {boxVisible ? (
           <Animated.View
             testID="moving-view"
