+++ packages/@react-native/tester/js/examples/FlatList/FlatList-onEndReached.js
@@ -9,20 +9,41 @@
  */
 
 'use strict';
-import {FlatList_onEndReached} from './FlatListExamples';
-const React = require('react');
+import type {RNTesterModuleExample} from '../../types/RNTesterTypes';
+import BaseFlatListExample from './BaseFlatListExample';
+import * as React from 'react';
 
-exports.title = 'FlatList onEndReached';
-exports.testTitle = 'Test onEndReached callback';
-exports.category = 'ListView';
-exports.documentationURL = 'https://reactnative.dev/docs/flatlist';
-exports.description =
-  'Scroll to end of list or tap Test button to see `onEndReached` triggered.';
-exports.examples = [
-  {
-    title: 'FlatList onEndReached',
-    render: function(): React.Element<typeof FlatList_onEndReached> {
-      return <FlatList_onEndReached />;
-    },
+export function FlatList_onEndReached(): React.Node {
+  const [output, setOutput] = React.useState('');
+  const exampleProps = {
+    onEndReached: info => setOutput('onEndReached'),
+    onEndReachedThreshold: 0,
+  };
+  const ref = React.useRef(null);
+
+  const onTest = () => {
+    const scrollResponder = ref?.current?.getScrollResponder();
+    if (scrollResponder != null) {
+      scrollResponder.scrollToEnd();
+    }
+  };
+
+  return (
+    <BaseFlatListExample
+      ref={ref}
+      exampleProps={exampleProps}
+      testOutput={output}
+      onTest={onTest}
+    />
+  );
+}
+
+export default ({
+  title: 'onEndReached',
+  name: 'onEndReached',
+  description:
+    'Scroll to end of list or tap Test button to see `onEndReached` triggered.',
+  render: function(): React.Element<typeof FlatList_onEndReached> {
+    return <FlatList_onEndReached />;
   },
-];
+}: RNTesterModuleExample);
