+++ packages/integration-test-app/package.json
@@ -13,7 +13,7 @@
     "@react-native-windows/automation-channel": "^0.0.3-10",
     "chai": "^4.2.0",
     "react": "17.0.2",
-    "react-native": "0.0.0-d40cb0e1b",
+    "react-native": "0.0.0-fa768df2a",
     "react-native-windows": "^0.0.0-canary.382"
   },
   "devDependencies": {
