+++ packages/playground/package.json
@@ -13,7 +13,7 @@
     "@react-native-picker/picker": "1.16.5",
     "@react-native-windows/tester": "0.0.1",
     "react": "17.0.2",
-    "react-native": "0.0.0-d40cb0e1b",
+    "react-native": "0.0.0-fa768df2a",
     "react-native-windows": "0.0.0-canary.382"
   },
   "devDependencies": {
