+++ packages/sample-apps/package.json
@@ -12,7 +12,7 @@
   },
   "dependencies": {
     "react": "17.0.2",
-    "react-native": "0.0.0-d40cb0e1b",
+    "react-native": "0.0.0-fa768df2a",
     "react-native-windows": "0.0.0-canary.382"
   },
   "devDependencies": {
